requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
qdrant-client==1.13.3
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
tqdm==4.66.1
markdown==3.5
beautifulsoup4==4.12.2
PyYAML==6.0.1
google-generativeai==0.3.1
backoff==2.2.1
python-dotenv==1.0.0